Document items added

+++ business-easy-core/src/main/java/com/businesseasy/core/repositories/ContactNoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ContactNoRepository extends JpaRepository<ContactNoEntity, Integer> {
+public interface ContactNoRepository extends JpaRepository<ContactNoEntity, Long> {
     Optional<ContactNoEntity> findByContactNo(String contactNo);
     List<ContactNoEntity> findAllByOwnerId(Long ownerId);
 }
